// src/types/User.ts
export interface UserData {
  nome: string;
  sobrenome: string;
  cpf: string;
  datanascimento: string;
  telefone: string;
  sexo: string;
  faixa: string;
  peso: string;
  ct: string;
  professor: string;
  cep: string;
  logradouro: string;
  bairro: string;
  cidade: string;
  estado: string;
  numero: string;
  complemento: string;
  email: string;
  senha: string;
  dica: string;
}